pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sie-autoppt"
version = "0.1.0"
description = "AI planning + deterministic PPTX rendering for enterprise slide generation"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "HUHU19961014" }
]
dependencies = [
  "python-pptx>=0.6.21",
  "openai>=1.0.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=4.9.0",
  "python-docx>=1.1.0",
  "pypdf>=4.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
sie-autoppt = "sie_autoppt.cli:main"

[tool.setuptools]
package-dir = {"" = "tools"}

[tool.setuptools.packages.find]
where = ["tools"]
include = ["sie_autoppt*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
